+++ packages/react/src/ErrorBoundaryGroup.spec.tsx
@@ -3,8 +3,9 @@ import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 import ms from 'ms'
 import { createElement } from 'react'
 import { describe, expect, it } from 'vitest'
+import { ErrorBoundary } from './ErrorBoundary'
+import { ErrorBoundaryGroup, useErrorBoundaryGroup } from './ErrorBoundaryGroup'
 import { useErrorBoundaryGroup_this_hook_should_be_called_in_ErrorBoundary_props_children } from './utils/assert'
-import { ErrorBoundary, ErrorBoundaryGroup, useErrorBoundaryGroup } from '.'
 
 const innerErrorBoundaryCount = 3
 const resetButtonText = 'reset button'
